backend/requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.37
botocore==1.35.37
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
soupsieve==2.6
urllib3==2.2.3
Werkzeug==3.0.4
